+++ src/pages/ManagementPlacePage/index.tsx
@@ -1,12 +1,12 @@
 import HeaderWithTitle from '@layouts/HeaderWithTitle';
 import MainLayout from '@layouts/MainLayout';
-import PlaceList from './components/PlaceList';
+import ManagementPlaceList from './components/ManagementPlaceList';
 
 const ManagementPlacePage = () => {
   return (
     <MainLayout headerType='both'>
       <HeaderWithTitle title='사업장 관리' />
-      <PlaceList />
+      <ManagementPlaceList />
     </MainLayout>
   );
 };
